[cudaaligner] Define test cases in separate compilation unit
